Removed save button on read-only view of configurations
