+++ base/uk.ac.stfc.isis.ibex.ui.configserver/src/uk/ac/stfc/isis/ibex/ui/configserver/commands/helpers/EditConfigHelper.java
@@ -20,30 +20,22 @@
 /**
  * 
  */
-package uk.ac.stfc.isis.ibex.ui.configserver.commands;
+package uk.ac.stfc.isis.ibex.ui.configserver.commands.helpers;
 
-import org.eclipse.jface.dialogs.MessageDialog;
 import org.eclipse.jface.window.Window;
 import org.eclipse.swt.widgets.Shell;
 
 import uk.ac.stfc.isis.ibex.configserver.ConfigServer;
 import uk.ac.stfc.isis.ibex.configserver.editing.EditableConfiguration;
-import uk.ac.stfc.isis.ibex.model.Awaited;
-import uk.ac.stfc.isis.ibex.model.UpdatedValue;
 import uk.ac.stfc.isis.ibex.ui.configserver.ConfigurationServerUI;
-import uk.ac.stfc.isis.ibex.ui.configserver.ConfigurationViewModels;
 import uk.ac.stfc.isis.ibex.ui.configserver.dialogs.EditConfigDialog;
 
 /**
  * A helper class to open config editing dialog boxes.
  */
-public class EditConfigHelper {
-    private static final String TITLE = "Edit Configuration";
-    private static final String SUB_TITLE_CURRENT = "Editing the current configuration";
+public class EditConfigHelper extends ConfigHelper {
 
-    private Shell shell;
     private ConfigServer server;
-    private ConfigurationViewModels configurationViewModels;
 
     /**
      * Constructor for the helper class.
@@ -55,45 +47,14 @@ public class EditConfigHelper {
         this.shell = shell;
         this.server = server;
         this.configurationViewModels = ConfigurationServerUI.getDefault().configurationViewModels();
+        
+        title = "Edit Configuration";
+        sub_title_current = "Editing the current configuration";
     }
 
-    /**
-     * Create a dialog box for editing the current config.
-     * 
-     * @param blockname The blockname we wish to edit, if null or empty we will
-     *            edit the whole config.
-     */
-    public void createDialogCurrent(String blockname) {
-        configurationViewModels.setModelAsCurrentConfig();
-        UpdatedValue<EditableConfiguration> config = configurationViewModels.getConfigModel();
-
-        if (Awaited.returnedValue(config, 1)) {
-            openDialog(SUB_TITLE_CURRENT, config.getValue(), configurationViewModels, blockname, true);
-        } else {
-            MessageDialog.openError(shell, "Error", "There is no current configuration, so it can not be edited.");
-        }
-    }
-
-    /**
-     * Create a dialog box for editing a config other than the current one.
-     * 
-     * @param configName The name of the config we wish to edit
-     */
-    public void createDialog(String configName) {
-        String subTitle = "Editing " + configName;
-
-        configurationViewModels.setModelAsConfig(configName);
-        UpdatedValue<EditableConfiguration> config = configurationViewModels.getConfigModel();
-
-        if (Awaited.returnedValue(config, 1)) {
-            openDialog(subTitle, config.getValue(), configurationViewModels, null, false);
-        }
-    }
-
-    private void openDialog(String subTitle, EditableConfiguration config,
-            ConfigurationViewModels configurationViewModels,
+    protected void openDialog(String subTitle, EditableConfiguration config,
             String blockname, boolean isCurrent) {
-        EditConfigDialog dialog = new EditConfigDialog(shell, TITLE, SUB_TITLE_CURRENT, config, false, false, blockname,
+        EditConfigDialog dialog = new EditConfigDialog(shell, title, sub_title_current, config, false, false, blockname,
                 configurationViewModels);
         if (dialog.open() == Window.OK) {
             if (dialog.doAsComponent()) {
