+++ base/uk.ac.stfc.isis.ibex.ui.configserver/src/uk/ac/stfc/isis/ibex/ui/configserver/dialogs/EditConfigDialog.java
@@ -20,59 +20,27 @@
 package uk.ac.stfc.isis.ibex.ui.configserver.dialogs;
 
 import java.util.Collection;
-import java.util.HashMap;
-import java.util.Map;
 
 import org.eclipse.jface.dialogs.IDialogConstants;
-import org.eclipse.jface.dialogs.TitleAreaDialog;
 import org.eclipse.jface.window.Window;
-import org.eclipse.swt.SWT;
 import org.eclipse.swt.events.SelectionAdapter;
 import org.eclipse.swt.events.SelectionEvent;
-import org.eclipse.swt.events.ShellAdapter;
-import org.eclipse.swt.events.ShellEvent;
-import org.eclipse.swt.graphics.Point;
-import org.eclipse.swt.layout.GridData;
 import org.eclipse.swt.widgets.Button;
 import org.eclipse.swt.widgets.Composite;
-import org.eclipse.swt.widgets.Control;
 import org.eclipse.swt.widgets.Shell;
 
-import com.google.common.base.Strings;
-
 import uk.ac.stfc.isis.ibex.configserver.Configurations;
-import uk.ac.stfc.isis.ibex.configserver.configuration.Configuration;
 import uk.ac.stfc.isis.ibex.configserver.editing.EditableConfiguration;
 import uk.ac.stfc.isis.ibex.ui.configserver.ConfigurationViewModels;
-import uk.ac.stfc.isis.ibex.ui.configserver.editing.ConfigEditorPanel;
-import uk.ac.stfc.isis.ibex.validators.MessageDisplayer;
 
 /**
  * Dialog for editing a configuration (top dialogue that contains save and save
  * as buttons).
  */
-public class EditConfigDialog extends TitleAreaDialog implements
-		MessageDisplayer {
+public class EditConfigDialog extends ConfigDetailsDialog {
 
-	private static final Point INITIAL_SIZE = new Point(650, 750);
-	private final String title;
-	private final String subTitle;
-
-    /** Error messages that are displayed. <Source, message> */
-	private Map<String, String> errorMessages = new HashMap<String, String>();
-
-	private EditableConfiguration config;
-
-	private ConfigEditorPanel editor;
-	private boolean doAsComponent = false;
-	private boolean isComponent;
-	private boolean isBlank;
-	private boolean firstShowing = true;
-    private String blockToEdit;
 	private Button saveAsBtn;
-
-    private ConfigurationViewModels configurationViewModels;
-
+	
     /**
      * @wbp.parser.constructor Constructor
      * 
@@ -86,14 +54,7 @@ public class EditConfigDialog extends TitleAreaDialog implements
      */
     public EditConfigDialog(Shell parentShell, String title, String subTitle, EditableConfiguration config,
             boolean isComponent, boolean isBlank, ConfigurationViewModels configurationViewModels) {
-		super(parentShell);
-		setShellStyle(getShellStyle() | SWT.DIALOG_TRIM | SWT.RESIZE);
-		this.title = title;
-		this.subTitle = subTitle;
-		this.config = config;
-		this.isComponent = isComponent;
-		this.isBlank = isBlank;
-        this.configurationViewModels = configurationViewModels;
+		super(parentShell, title, subTitle, config, isComponent, isBlank, configurationViewModels);
 	}
 
     /**
@@ -110,41 +71,8 @@ public class EditConfigDialog extends TitleAreaDialog implements
      */
     public EditConfigDialog(Shell parentShell, String title, String subTitle, EditableConfiguration config,
             boolean isComponent, boolean isBlank, String blockName, ConfigurationViewModels configurationViewModels) {
-        this(parentShell, title, subTitle, config, isComponent, isBlank, configurationViewModels);
-        this.blockToEdit = blockName;
-        
+        super(parentShell, title, subTitle, config, isComponent, isBlank, blockName, configurationViewModels);
     }
-    
-    private void showBlockDialog() {
-    	// Open the edit block dialog on shellActivated as must be done after the Config Editor dialog is shown.
-        getShell().addShellListener(new ShellAdapter() {
-			@Override
-			public void shellActivated(ShellEvent e) {
-				// Only open when the shell is activated for the first time
-				if (firstShowing) {
-					Shell s = (Shell) e.getSource();
-                    s.setVisible(true);
-		            openBlocksTab();
-		            openEditBlockDialog();
-		            firstShowing = false;
-				}
-			}
-		});    	
-    }
-    
-	@Override
-	protected Control createDialogArea(Composite parent) {
-        editor = new ConfigEditorPanel(parent, SWT.NONE, this, isComponent, configurationViewModels);
-		editor.setLayoutData(new GridData(SWT.FILL, SWT.FILL, true, true, 1, 1));
-		setTitle(subTitle);
-		editor.setConfigToEdit(config);
-		
-        if (!Strings.isNullOrEmpty(blockToEdit)) {
-        	showBlockDialog();
-        }
-		
-		return editor;
-	}
 
 	@Override
 	protected void createButtonsForButtonBar(Composite parent) {
@@ -181,49 +109,19 @@ public class EditConfigDialog extends TitleAreaDialog implements
 		createButton(parent, IDialogConstants.CANCEL_ID, "Cancel", false);
 
         // Show any error message and set buttons after creating those buttons
-        this.showErrorMessage();
+        super.showErrorMessage();
 	}
 
 	@Override
-	protected void configureShell(Shell shell) {
-		super.configureShell(shell);
-		shell.setText(title);
-	}
-
-	@Override
-	protected Point getInitialSize() {
-		return INITIAL_SIZE;
-	}
-
-	@Override
-	public void setErrorMessage(String source, String error) {
-		errorMessages.put(source, error);
-		showErrorMessage();
-        openEditBlockDialog();
-	}
-
-    /**
-     * Show the current error messages.
-     */
-	private void showErrorMessage() {
-		StringBuilder sb = new StringBuilder();
-		for (String key : errorMessages.keySet()) {
-			if (errorMessages.get(key) != null) {
-				sb.append(errorMessages.get(key));
-				sb.append("  ");
-			}
-		}
-
-		if (sb.length() > 0) {
-			setErrorMessage(sb.toString());
-			// Don't allow save until errors are cleared
-			setOKEnabled(false);
-			saveAsBtn.setEnabled(false);
-		} else {
-			setErrorMessage(null);
+	public void setErrorMessage(String newErrorMessage) {
+		if (newErrorMessage == null) {
 			setOKEnabled(true);
 			saveAsBtn.setEnabled(true);
+		} else {
+			setOKEnabled(false);
+			saveAsBtn.setEnabled(false);
 		}
+		super.setErrorMessage(newErrorMessage);
 	}
 
 	private void setOKEnabled(boolean value) {
@@ -232,42 +130,4 @@ public class EditConfigDialog extends TitleAreaDialog implements
 			okButton.setEnabled(value);
 		}
 	}
-
-    /**
-     * 
-     * @return the configuration
-     */
-	public Configuration getConfig() {
-		return config.asConfiguration();
-	}
-
-    /**
-     * 
-     * @return the component being edited
-     */
-	public Configuration getComponent() {
-		return config.asComponent();
-	}
-
-    /**
-     * 
-     * @return whether dialogue is for editing a component
-     */
-	public boolean doAsComponent() {
-		return doAsComponent;
-	}
-
-    /**
-     * Open tab for blocks.
-     */
-    public void openBlocksTab() {
-        editor.openTab(ConfigEditorPanel.BLOCK_TAB_NAME);
-    }
-
-    /**
-     * Open edit block dialogue.
-     */
-    public void openEditBlockDialog() {
-        editor.openEditBlockDialog(this.blockToEdit);
-    }
 }
